+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.51.0
 
 * Allow setting the following configurations with the listener prefix (e.g. `listener.name.listener1-9900.`): `connections.max.reauth.ms`, `max.connections*` and `max.connection.creation.rate`.
+* The `ServerSideApplyPhase1` feature gate moves to beta stage and is enabled by default.
+  If needed, `ServerSideApplyPhase1` can be disabled in the feature gates configuration in the Cluster Operator.
 
 ### Major changes, deprecations, and removals